refactor code to mui and moved options and data to utils
