--- src/components/common/ui/column-chart/ColumnChart.module.scss
@@ -1,14 +1,0 @@
-.chartContainer {
-  height: 314px;
-  max-width: 404px;
-  border: 1px solid #404040;
-  border-radius: 8px;
-  padding: 10px;
-}
-
-.title {
-  font-size: 16px;
-  color: #E5E5E5;
-  text-align: center;
-  margin-bottom: 15px;
-}
