+++ src/components/common/ui/column-chart/ColumnChart.styles.ts
@@ -1,0 +1,9 @@
+import { SxProps, Theme } from '@mui/material/styles';
+
+export const chartContainer: SxProps<Theme> = {
+  height: 314,
+  maxWidth: 404,
+  border: '1px solid #404040',
+  borderRadius: 8,
+  padding: 10,
+};
